feat(transcription): add setup onboarding route
