+++ apps/whispering/src/lib/components/settings/index.ts
@@ -11,3 +11,4 @@ export { default as RecordingModeSelector } from './selectors/RecordingModeSelec
 export { default as TranscriptionSelector } from './selectors/TranscriptionSelector.svelte';
 export { default as TransformationSelector } from './selectors/TransformationSelector.svelte';
 export { default as VadDeviceSelector } from './selectors/VadDeviceSelector.svelte';
+export { default as TranscriptionRuntimeSetup } from './TranscriptionRuntimeSetup.svelte';
