+++ apps/whispering/src/routes/(app)/_layout-utils/register-onboarding.ts
@@ -1,46 +1,22 @@
 import { goto } from '$app/navigation';
 import { report } from '$lib/report';
-import {
-	getSelectedTranscriptionService,
-	isTranscriptionServiceConfigured,
-} from '$lib/settings/transcription-validation';
+import { getTranscriptionSetupReadiness } from '$lib/settings/transcription-validation';
 
 /**
- * Checks if the user has configured the necessary API keys/settings for their selected transcription service.
- * Shows an onboarding toast if configuration is missing.
+ * Checks the current setup facts and points incomplete devices to setup.
  */
 export function registerOnboarding() {
-	const selectedService = getSelectedTranscriptionService();
+	if (window.location.pathname.startsWith('/setup')) return;
 
-	// Check transcription service configuration
-	if (!selectedService) {
-		report.info({
-			title: 'Welcome to Whispering!',
-			description: 'Please select a transcription service to get started.',
-			action: {
-				label: 'Configure',
-				onClick: () => goto('/settings/transcription'),
-			},
-		});
-		return;
-	}
+	const readiness = getTranscriptionSetupReadiness();
+	if (readiness.isReady) return;
 
-	if (!isTranscriptionServiceConfigured(selectedService)) {
-		const description = (
-			{
-				cloud: `Please add your ${selectedService.label} API key to get started.`,
-				'self-hosted': `Please set your ${selectedService.label} server URL to get started.`,
-				local: `Please download or select a ${selectedService.label} model to get started.`,
-			} as const
-		)[selectedService.location];
-
-		report.info({
-			title: 'Welcome to Whispering!',
-			description,
-			action: {
-				label: 'Set up',
-				onClick: () => goto('/settings/transcription'),
-			},
-		});
-	}
+	report.info({
+		title: 'Finish Whispering setup',
+		description: readiness.primaryIssue ?? 'Complete setup before recording.',
+		action: {
+			label: 'Open setup',
+			onClick: () => goto('/setup'),
+		},
+	});
 }
